Move non-widget selection logic to `util.selection`
